+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [3.58.12] - 2026-05-07
+
+### ♻️ Refactoring
+
+- **reports**: Dynamic Report Builder - Replace 8 static screens (#377)
+
+## [3.58.11] - 2026-05-05
+
+### 🐛 Bug Fixes
+
+- Priority bug fixes and quality improvements (#375)
+
 ## [3.58.9] - 2026-05-04
 
 ### 🐛 Bug Fixes